# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway Service
  description: REST API для проксирования запросов и аутентификации
  version: 1.0.0
servers:
  - url: http://localhost:5000

paths:
  /health:
    get:
      tags:
        - Healthcheck
      summary: Проверка работоспособности сервиса
      responses:
        '200':
          description: Сервис работает нормально
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  /users/v1/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
                - email
              properties:
                login:
                  type: string
                  minLength: 3
                password:
                  type: string
                  format: password
                  minLength: 8
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /users/v1/login:
    post:
      tags:
        - Authentication
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
              properties:
                login:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /users/v1:
    get:
      tags:
        - User Profile
      summary: Получение профиля пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Данные профиля
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  login:
                    type: string
                  email:
                    type: string
                  first_name:
                    type: string
                  last_name:
                    type: string
                  birth_date:
                    type: string
                    format: date
                  phone_number:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '401':
          description: Невалидный токен
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    put:
      tags:
        - User Profile
      summary: Обновление профиля пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                birth_date:
                  type: string
                  format: date
                email:
                  type: string
                  format: email
                phone_number:
                  type: string
      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

securitySchemes:
  BearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT
    description: JWT токен авторизации